[babylon] Create State#with(newState, callback)
This method runs some parser code with a given state,
and resets the state to the original after that the
callback is run.
It returns the value returned by the callback.

Example:

  node.body = this.state.with(
    { inFunction: false, labels: [] },
    () => this.parseBlock(false),
  );

It avoids manually caching the state each time, like
it was needed until now:

  const oldInFunction = this.state.inFunction;
  const oldLabels = this.state.labels;
  this.state.inFunction = false;
  this.state.labels = [];
  node.body = this.parseBlock(false);
  this.state.inFunction = oldInFunction;
  this.state.labels = oldLabels;

Possible future expansion:
When this method wraps the whole body of a method
(which happens quite often), it could be replaced
by a decorator. e.g.

  @State.with({ inParameters: true }
  parseFunctionParams(...): void {
    this.expect(tt.parenL);
    node.params = this.parseBindingList(...);
  }